+++ spec/rubocop/cli/cli_autocorrect_spec.rb
@@ -1169,13 +1169,13 @@ describe RuboCop::CLI, :isolated_environment do
     expect(IO.read('example.rb')).to eq("some_method(a)\n")
   end
 
-  it 'does not hang SpaceAfterPunctuation and SpaceInsideBrackets' do
+  it 'does not hang SpaceAfterPunctuation' do
     create_file('example.rb', 'puts [1, ]')
     Timeout.timeout(10) do
       expect(cli.run(%w[--auto-correct])).to eq(0)
     end
     expect($stderr.string).to eq('')
-    expect(IO.read('example.rb')).to eq("puts [1]\n")
+    expect(IO.read('example.rb')).to eq("puts [1 ]\n")
   end
 
   it 'can be disabled for any cop in configuration' do